pyproject.toml
[tool.poetry]
name = "qa-guru-diploma"
version = "0.1.0"
description = ""
authors = ["A-d-a-m <adam3928@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
selene = "2.0.0rc4"
allure-pytest = "^2.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

